requirements.txt
holidays
google-cloud-firestore
google-cloud-storage
google-genai
google-adk
python-dotenv